Some refactoring, renaming
